proposal_backend/requirements.txt
flask==2.3.2
flask-pymongo==2.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
bcrypt==4.0.1
pyjwt==2.8.0
python-docx==0.8.11
# gridfs==0.6.2
pymongo==4.12.1